timedata: Removed initial value from CMedianFilter
Reason: Not really used and it caused sync issues when only connected to 
one peer that is way higher than your chain because it calculates median 
between 0 and "connected peer height". This lead to not accepting the 
block headers of the peer because in this case it assumed unintentionaly 
that the peer is lower than us.